move specs/acceptance/server to scripts/test-server
