+++ tutor/scripts/test-server/backend/dashboard.js
@@ -1,6 +1,6 @@
 const Factory = require('object-factory-bot');
-require('../../../factories/dashboard');
-require ('../../../factories/research_survey');
+require('../../../specs/factories/dashboard');
+require ('../../../specs/factories/research_survey');
 const { times, pick, clone, find } = require('lodash');
 const { data: bootstrapData } = require('./bootstrap');
 const { now } = require('../time-now');
